+++ src/Event/WebView/AfterRender.php
@@ -8,14 +8,13 @@ use Yiisoft\View\Event\AfterRenderEventInterface;
 use Yiisoft\View\WebView;
 
 /**
- * `AfterRender` event is triggered by {@see WebView::renderFile()} right after it renders a view file.
+ * `AfterRender` event is triggered by {@see WebView::render()} right after it renders a view file.
  */
 final class AfterRender implements AfterRenderEventInterface
 {
     /**
      * @param string $file The view file being rendered.
-     * @param array $parameters The parameters array passed to the {@see WebView::render()}
-     * or {@see WebView::renderFile()} method.
+     * @param array $parameters The parameters array passed to the {@see WebView::render()} method.
      */
     public function __construct(
         private WebView $view,