+++ src/ViewInterface.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Yiisoft\View;
 
 use InvalidArgumentException;
-use RuntimeException;
+use LogicException;
 use Throwable;
 use Yiisoft\View\Exception\ViewNotFoundException;
 
@@ -59,14 +59,14 @@ interface ViewInterface
     /**
      * Returns a new instance with the specified view context instance.
      *
-     * @param ViewContextInterface $context The context under which the {@see renderFile()} method is being invoked.
+     * @param ViewContextInterface $context The context under which the {@see render()} method is being invoked.
      */
     public function withContext(ViewContextInterface $context): static;
 
     /**
      * Returns a new instance with the specified view context path.
      *
-     * @param string $path The context path under which the {@see renderFile()} method is being invoked.
+     * @param string $path The context path under which the {@see render()} method is being invoked.
      */
     public function withContextPath(string $path): static;
 
@@ -249,45 +249,25 @@ interface ViewInterface
      *
      * The view to be rendered can be specified in one of the following formats:
      *
-     * - The name of the view starting with a slash to join the base path {@see getBasePath()} (e.g. "/site/index").
-     * - The name of the view without the starting slash (e.g. "site/index"). The corresponding view file will be
+     * - the absolute path to the view file, e.g. "/path/to/view.php";
+     * - the name of the view starting with `//` to join the base path {@see getBasePath()}, e.g. "//site/index";
+     * - the name of the view starting with `./` to join the directory containing the view currently being rendered
+     *   (i.e., this happens when rendering a view within another view), e.g. "./widget";
+     * - the name of the view without the starting `//` or `./` (e.g. "site/index"). The corresponding view file will be
      *   looked for under the {@see ViewContextInterface::getViewPath()} of the context set via {@see withContext()}.
-     *   If the context instance was not set {@see withContext()}, it will be looked for under the directory containing
-     *   the view currently being rendered (i.e., this happens when rendering a view within another view).
+     *   If the context instance was not set {@see withContext()}, it will be looked for under the base path.
      *
      * @param string $view The view name.
      * @param array $parameters The parameters (name-value pairs) that will be extracted and made available in the view
      * file.
      *
-     * @throws RuntimeException If the view cannot be resolved.
+     * @throws LogicException If the view cannot be resolved.
      * @throws ViewNotFoundException If the view file does not exist.
      * @throws Throwable
-     *
-     * {@see renderFile()}
      *
      * @return string The rendering result.
      */
     public function render(string $view, array $parameters = []): string;
-
-    /**
-     * Renders a view file.
-     *
-     * If the theme was set {@see setTheme()}, it will try to render the themed version of the view file
-     * as long as it is available.
-     *
-     * If the renderer was set {@see withRenderers()}, the method will use it to render the view file. Otherwise,
-     * it will simply include the view file as a normal PHP file, capture its output and return it as a string.
-     *
-     * @param string $viewFile The full absolute path of the view file.
-     * @param array $parameters The parameters (name-value pairs) that will be extracted and made available in the view
-     * file.
-     *
-     * @throws Throwable
-     * @throws ViewNotFoundException If the view file does not exist
-     *
-     * @return string The rendering result.
-     */
-    public function renderFile(string $viewFile, array $parameters = []): string;
 
     /**
      * Returns the localized version of a specified file.